serveral changes
Changed they way images are displayed in the product details, now images are stored in a doubly linked list that acts the local database for the product images, allowing store owners to add, change and delete images and seeing the update without having to refresh the page. Images are now displayed as a carousel displaying only one image at the time. The user can change between images by pressing the arrow buttons at each side of the biggest image or by clicking on the images previews.
*favicon actualized too